Add transform that adds the @key directives to our Node types
Before this we were relying on not visiting all nodes for pruning the schema, I
added some extra code to remove unreachable types. This uncovered a bug in our
existing test, so I fixed those tests as well.

During my changes here I triggered some strage behaviour related to running the
Federation transforms multiple times when `removeFederationDefinitions` is
false. Because of this I made the code skip this step if the _Entity union type
already exists.